requirements.txt
holidays==0.47
requests==2.31.0
urllib3==2.2.1
python-dotenv==1.0.1